Refactor MCO2 code
Fix trainer functions due to issues with saving. Also,
refactored trainer to accomodate sizeability and flexibility
due to nested arrayLists making it hard to work with
previously.

Update feature for Pokemon class, where it used to take String
moves and item. Now takes Item and Move class, items and move set
accordingly.

Refactor the rest of the code accordingly, due to major changes to
Pokemon class and function class.

Documentation is lacking, to be fixed in future build pushes