Fix evolution feature and unique pokemon issues
I added a constructor in Pokemon.java that allows you
to pass an object of itself and it makes that copy. This
is because while I was testing, it seemed like the
evolution passed for ALL vulpix copies (yes even the one in the
databse). So, I figured that I fix that issue by fixing
the unique copy of the Pokemon.

Im unsure if theres any more cases where the pokemon should
be copied or so, a double review of the code wouldn't hurt.
Though in the next commit expect SOME GUI implementations.
That is all.